completed basic working pipeline

+++ JastraSpecification/src/main/java/io/github/inbanithi/jastra/specification/core/ControlInstruction.java
@@ -2,8 +2,8 @@ package io.github.inbanithi.jastra.specification.core;
 
 public class ControlInstruction {
 
-    public static byte ENTRY = (byte) 0xB0;
+    public static int ENTRY = 0xB0;
 
-    public static byte HALT = (byte) 0xB7;
+    public static int HALT = 0xB7;
 
 }
